tests/requirements.txt
# Test requirements for python-selve

# Core dependencies
pyserial>=3.5
pyserial-asyncio>=0.6
untangle>=1.2.0

# Test framework
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-mock>=3.10.0
pytest-cov>=4.0.0
pytest-html>=3.1.0
pytest-xdist>=3.0.0

# Utilities
psutil>=5.9.0
coverage>=7.0.0
colorama>=0.4.0

# Project dependencies needed for testing
pyserial>=3.5
pyserial-asyncio>=0.6
untangle>=1.2.1
